+++ R/whirl_r_session.R
@@ -14,7 +14,6 @@ whirl_r_session <- R6::R6Class(
     #' @return A [whirl_r_session] object
     initialize = \(
       # jscpd:ignore-start
-      verbosity_level = zephyr::get_verbosity_level("whirl"),
       check_renv = zephyr::get_option("check_renv", "whirl"),
       track_files = zephyr::get_option("track_files", "whirl"),
       out_formats = zephyr::get_option("out_formats", "whirl"),
@@ -29,7 +28,6 @@ whirl_r_session <- R6::R6Class(
       # jscpd:ignore-end
     ) {
       wrs_initialize(
-        verbosity_level,
         check_renv,
         track_files,
         out_formats,
@@ -70,20 +68,6 @@ whirl_r_session <- R6::R6Class(
       wrs_check_status(self, private, super)
     },
 
-    #' @description Update the progress bar
-    #' @param ... parsed to [cli::cli_progress_update()]
-    #' @return [invisible] self
-    pb_update = \(...) {
-      wrs_pb_update(..., self = self, private = private, super = super)
-    },
-
-    #' @description Finalise the progress bar
-    #' @param status Status of the script. success, warning, or error
-    #' @return [invisible] self
-    pb_done = \(status) {
-      wrs_pb_done(status, self, private, super)
-    },
-
     #' @description Log the script
     #' @param script The script to log
     #' @return [invisible] self
@@ -97,21 +81,15 @@ whirl_r_session <- R6::R6Class(
       wrs_create_log(self, private, super)
     },
 
-    #' @description Finish the log
-    #' @return [invisible] self
-    log_finish = \() {
-      wrs_log_finish(self, private, super)
-    },
-
-    #' @description Create all log outputs
+    #' @description Finish the log and create outputs
     #' @param out_dir [character] Output directory for the log
     #' @param format [character] Output formats to create
-    #' @return [invisible],[list] of logging information
-    create_outputs = \(
+    #' @return [invisible] self
+    log_finish = \(
       out_dir,
       format = zephyr::get_option("out_formats", "whirl")
     ) {
-      wrs_create_outputs(out_dir, format, self, private, super)
+      wrs_log_finish(out_dir, format, self, private, super)
     }
   ),
   private = list(
@@ -119,7 +97,6 @@ whirl_r_session <- R6::R6Class(
     finalize = \() {
       wrs_finalize(self, private, super)
     },
-    verbosity_level = NULL,
     wd = NULL,
     track_files = NULL,
     out_formats = NULL,
@@ -128,7 +105,6 @@ whirl_r_session <- R6::R6Class(
     track_files_keep = NULL,
     approved_packages = NULL,
     check_renv = NULL,
-    pb = NULL,
     current_script = NULL,
     track_files_log = "log_msg.json",
     result = NULL
@@ -137,7 +113,6 @@ whirl_r_session <- R6::R6Class(
 )
 
 wrs_initialize <- function(
-  verbosity_level,
   check_renv,
   track_files,
   out_formats,
@@ -153,7 +128,6 @@ wrs_initialize <- function(
   super$initialize(wait_timeout = wait_timeout) # uses callr::r_session$initialize()
 
   private$wd <- withr::local_tempdir(clean = FALSE)
-  private$verbosity_level <- verbosity_level
   private$check_renv <- check_renv
   private$track_files <- track_files
   private$out_formats <- out_formats
@@ -161,12 +135,6 @@ wrs_initialize <- function(
   private$track_files_keep <- track_files_keep
   private$approved_packages <- approved_packages
   private$log_dir <- log_dir
-
-  # If the stream does not support dynamic tty, which is needed for progress
-  # bars to update in place, the verbosity is downgraded.
-  if (private$verbosity_level == "verbose" && !cli::is_dynamic_tty()) {
-    private$verbosity_level <- "minimal"
-  }
 
   super$run(func = setwd, args = list(dir = private$wd))
 
@@ -219,8 +187,7 @@ wrs_finalize <- function(self, private, super) {
 wrs_print <- function(self, private, super) {
   msg <- c(
     utils::capture.output(super$print()),
-    "Working Directory: {private$wd}",
-    "Verbose: {private$verbosity_level}"
+    "Working Directory: {private$wd}"
   )
 
   cli::cli_bullets(
@@ -233,25 +200,8 @@ wrs_print <- function(self, private, super) {
   return(invisible(self))
 }
 
-wrs_pb_update <- function(..., self, private, super) {
-  if (!is.null(private$pb)) {
-    private$pb$update(...)
-  }
-  return(invisible(self))
-}
-
-wrs_pb_done <- function(status, self, private, super) {
-  if (!is.null(private$pb)) {
-    private$pb$done(status)
-  }
-  return(invisible(self))
-}
-
 wrs_poll <- function(timeout, self, private, super) {
   status <- super$poll_process(timeout)
-  if (status == "timeout") {
-    self$pb_update()
-  }
   return(status)
 }
 
@@ -284,6 +234,7 @@ wrs_check_status <- function(self, private, super) {
 }
 
 wrs_log_script <- function(script, self, private, super) {
+  zephyr::msg_debug("Running script {.file {script}}")
   private$current_script <- script
 
   saveRDS(
@@ -323,16 +274,6 @@ wrs_log_script <- function(script, self, private, super) {
     )
   }
 
-  # Execute the script
-  if (private$verbosity_level != "quiet") {
-    private$pb <- pb_script$new(
-      script = private$current_script,
-      use_progress = private$verbosity_level == "verbose"
-    )
-  }
-
-  self$pb_update(status = "Running script")
-
   self$call(
     func = \(...) quarto::quarto_render(...),
     args = list(
@@ -353,7 +294,7 @@ wrs_log_script <- function(script, self, private, super) {
 }
 
 wrs_create_log <- function(self, private, super) {
-  self$pb_update(status = "Creating log")
+  zephyr::msg_debug("Creating log for {.file {private$current_script}}")
 
   if (private$track_files) {
     strace_msg <- private$wd |>
@@ -391,16 +332,49 @@ wrs_create_log <- function(self, private, super) {
   return(invisible(self))
 }
 
-wrs_log_finish <- function(self, private, super) {
+wrs_log_finish <- function(out_dir, format, self, private, super) {
   private$result <- private$wd |>
     file.path("result.rds") |>
     readRDS()
 
-  if (!is.null(private$pb)) {
-    self$pb_done(status = private$result[["status"]][["message"]])
-  }
+  output <- wrs_create_outputs(out_dir, format, self, private, super)
 
-  return(invisible(self))
+  wrs_report_status(
+    status = output$status$message,
+    script = private$current_script,
+    logs = output$logs
+  )
+
+  invisible(output)
+}
+
+wrs_report_status <- function(status, script, logs) {
+  script_msg <- create_cli_links(
+    text = basename(script),
+    href = script
+  )
+
+  logs_msg <- create_cli_links(
+    text = get_file_ext(logs),
+    href = logs
+  )
+
+  switch(
+    EXPR = status,
+    success = zephyr::msg_success(
+      "{script_msg}: Completed succesfully. See {cli::qty(logs_msg)}log{?s} {logs_msg}."
+    ),
+    warning = zephyr::msg(
+      "{script_msg}: Completed with warnings. See {cli::qty(logs_msg)}log{?s} {logs_msg}.",
+      msg_fun = cli::cli_alert_warning # Since zephyr::msg_warning only shows when verbose
+    ),
+    error = zephyr::msg_danger(
+      "{script_msg}: Completed with errors. See {cli::qty(logs_msg)}log{?s} {logs_msg}."
+    ),
+    cli::cli_abort(
+      "{script}: Completed with unknown status {.emph {status}}. See {cli::qty(logs_msg)}log{?s} {logs_msg}."
+    )
+  )
 }
 
 wrs_create_outputs <- function(out_dir, format, self, private, super) {